+++ src/merrypopins/__init__.py
@@ -11,7 +11,7 @@ A modular pipeline for nanoindentation analysis. Includes tools for:
  - `make_dataset`: Construct enriched datasets by running the full pipeline and exporting annotated results and visualizations.
 """
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"
 
 # Expose submodules at the package level
 from . import load_datasets, preprocess, locate, statistics, make_dataset
