+++ Dockerfile
@@ -1,5 +1,5 @@
 # Use Python as the base image
-FROM python:3.13-slim
+FROM python:3.14-slim
 
 # Set working directory
 WORKDIR /app
