### 2026-02-04 #9933 @editorialbot
[REVIEW]: `merrypopins`: A Python package for nanoindentation data science
**Submitting author:** @mnky9800n (<a href="http://orcid.org/0000-0003-2258-3836">John M. Aiken</a>)
**Repository:** https://github.com/SerpRateAI/merrypopins
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.3
