### 2026-02-05 #9933 @KennethKwabenaKenney
@Fei-Tao
Is is compulsory for the authors to create specific sections for State of Field, Software design and Research Impact Statement ?
Because in this case where authors puts all this sections into the Statement of need, will that be considered?

Secondly, If AI usage is not disclosed, should that be querried?
