### 2026-02-05 #9933 @Fei-Tao
Hi @KennethKwabenaKenney, thanks for your prompt start of the review.
1. Is is compulsory for the authors to create specific sections for State of Field, Software design and Research Impact Statement ?
JOSS does not requires author to state each of these sections. It is acceptable to state these points in a single section.

2. Secondly, If AI usage is not disclosed, should that be querried?
JOSS allows AI assistance. Regardless of whether an author leveraged AI assistance, they need to show the human work: the problem framing, key design decisions and abstractions, and the practices that make the software usable and sustainable for others.

Feel free to let me know if you have any other questions or concerns.
