### 2026-02-08 #9933 @Manishms18
@Fei-Tao Overall, the software is well designed, reproducible, and follows strong open-source practices. The repository includes tests, documentation, tutorials, and sample datasets, and the software addresses a meaningful need in nano-indentation data analysis. I recommend acceptance after minor revisions.

Required Improvements

State of the Field
Provide clearer comparison with existing indentation analysis tools or workflows.
Clarify the unique contributions and justify building a new framework instead of extending existing solutions.

AI Usage Disclosure
Add a clearly labeled AI usage disclosure section.
Explicitly state whether generative AI tools were used during software or manuscript development.

Recommended Improvements (Optional)
Provide additional evidence of research impact or external adoption if available.
Add minor documentation enhancements such as workflow diagrams or quick method comparison tables.
Include clearer references to reproducibility materials (tests, datasets, tutorials) within the paper.
