### 2026-02-09 #9933 @KennethKwabenaKenney
Hello Editor @Fei-Tao

**SUMMARY**
This paper presents merrypopins, for automated detection and statistical analysis of pop-in events in nanoindentation experiments. The software addresses an important and timely need in nano-mechanics and materials science by replacing manual, subjective pop-in identification with reproducible, automated methods. The manuscript is generally well written, clearly motivated, and aligned with the JOSS scope and expectations.
Also the Streamlit app is impressive and excellent
The contribution is scientifically relevant, and the software is openly available. I Just have a few clarifications and minor improvements

---

**MAJOR COMMENTS**
1. The Statement of Need is well articulated and convincingly motivates why automated pop-in detection is necessary. The connection between pop-ins, dislocation avalanches, and earthquake analogues is particularly compelling. While the authors note that most prior approaches rely on manual inspection or private scripts, it would strengthen the paper to briefly clarify:
- Whether any existing open-source tools partially address pop-in detection
- What specifically distinguishes merrypopins (e.g., multi-method detection framework, extensibility, no-code interface, statistical suite)

2. Software description: It will be great to provide a brief overview of how the different detection methods are selected or combined in practice

3. The inclusion of Isolation Forests and convolutional autoencoders is a strength. However, it will be great to briefly state whether these models are trained by the user or provided preconfigured although this is seen in the software.
4. Reproducibility and testing: While JOSS does not require extensive benchmarking, it would be helpful to briefly mention how detection results can be validated

---

**MINOR COMMENTS & SUGGESTIONS**
- Line 29: The statement “merrypopins marks the first attempt to automate pop-in detection” is strong; consider softening or -qualifying unless this claim can be clearly supported.
- Lines 56–57: Minor typographical errors (“experimetal”, “merrypopins primary uses”) should be corrected.
- Add the AI usage disclosure

---

**RECOMMENDATION**

I recommend acceptance pending minor revisions. The requested changes are primarily clarifications and small improvements that will enhance clarity, reproducibility, and alignment with JOSS expectations, without requiring substantial restructuring of the manuscript or software.

Overall, merrypopins is a valuable contribution to the open-source ecosystem for nanoindentation data analysis and will be of interest to researchers across materials science, geophysics, and nano-mechanics.
