### 2026-02-13 #9933 @Manishms18
@Fei-Tao The repository demonstrates strong software engineering practices and satisfies JOSS code review criteria. The software appears technically sound and ready for publication, subject to the minor manuscript revisions noted earlier, address in issue section.
