### 2026-02-17 #9933 @KennethKwabenaKenney
@Fei-Tao
I have completed my review of the Merrypopins repository codes.

The project is clearly organized, properly licensed, and includes sufficient documentation and instructions for installation and use. The software implementation aligns with the claims made in the manuscript.

Overall, the repository satisfies the JOSS guidelines, and I have no major concerns.
