+++ .github/workflows/python-app.yml
@@ -44,7 +44,7 @@ jobs:
           pytest --cov=merrypopins --cov-report=xml
 
       - name: 📤 Upload coverage to Codecov
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v6
         with:
           files: ./coverage.xml
           token: ${{ secrets.CODECOV_TOKEN }}
