### 2026-08-25 #9933 @mnky9800n
Do I need to tell editorial bot to accept? I assumed you would need to do it since I’m the author.
