### 2026-08-27 #9933 @rkurchin
@mnky9800n, your paper is also missing a Research Impact Statement, so that needs to be added. Even without it the paper is a bit long and this will obviously extend it further (JOSS targets 1750 words or fewer), so please also compress things a bit.
