### 2026-08-27 #9933 @rkurchin
I'm going to start working on the rest of my final checklist while we wait to hear back...note that some of the copyediting points may or may not remain relevant as you compress the manuscript.

- [ ] Please remove the `verbatim` text styling in the paper title
- [ ] Cooley reference: please ensure {Capitalization} of Fourier
- [ ] first Aiken reference: Similarly, Substack should be capitalized
- [ ] In Summary, sentence beginning with "Experiments performed with a spherical tip...": broken parallelism please rephrase second portion to "...and, by extension, determine the yield point, defining the transition..."
- [ ] Statement of need, start of second paragraph: I think there's a word missing, should probably read "Pop-ins are linked to dislocation _motion_..." or similar
- [ ] last paragraph of statement of need: the name of the field is "materials science" (with an S)
- [ ] State of the field: you mention three groups and then never enumerate them. I would suggest either creating a numbered list or just adding "First, ..." etc. before each type
- [ ] same section: hyphenate "closed-source"
- [ ] Fourth paragraph of State of the Field: correct to "...and plots Weibull or time- and temperature-dependent..." (i.e. just add hyphens)
- [ ] same paragraph: again, you mention two things and then don't call them out. Here you can probably solve it by just adding a colon after "target audience" since they're both in the same sentence
- [ ] last paragraph of State of the field, last sentence: hyphenate "MIT-licensed"
- [ ] Software design: hyphenate "end-to-end"
