### 2026-08-28 #9933 @mnky9800n
@rkurchin thanks for the careful read. Everything on your list is done on the `JOSS` branch.

I have added the Research Impact Statement as its own section. I kept it to things that can be checked against the repository rather than claiming outside adoption we cannot document yet: the twelve tagged releases and the Zenodo archive, the 36 real indentation curves shipped with their metadata and overlay plots, the five tutorial notebooks, the test suite covering all five modules including synthetic curves with pop-ins of known position and size, the Streamlit app, and the SerpRateAI and EPOS-eNLarge work the package supports.

On length, the paper is now 1707 words, down from 1888, so it is under 1750 even with the new section. Most of that came from tightening the size-effect discussion in the Statement of need and the detector descriptions in Software design. The rest came from moving the reproducibility evidence out of "Validating detections" and into the new section instead of saying it twice.

Your copyediting points, in order:

- [x] Removed the verbatim styling from the title. It now reads "merrypopins: A Python package for nanoindentation data science", which also matches the title in the Zenodo archive.
- [x] Cooley reference: Fourier is now braced so the capital survives BibTeX.
- [x] First Aiken reference: Substack is now capitalised.
- [x] Summary: "...and, by extension, determine the yield point, defining the transition from elastic to plastic deformation".
- [x] Statement of need: "Pop-ins are linked to dislocation motion in crystalline materials".
- [x] "materials science".
- [x] State of the field: the three groups are now called out as "First, vendor software...", "Second, open-source indentation packages...", "Third, a few tools address pop-ins directly...". The PopIn toolbox and the Kossman & Bigerelle classifier both sit in that third group, which also let me drop a couple of transition sentences.
- [x] "closed-source".
- [x] "plots Weibull or time- and temperature-dependent cumulative distributions".
- [x] Merged with a colon: "Two things make it a poor foundation for our target audience: it requires a commercial MATLAB licence, and it is scoped to the first and second pop-in of a curve rather than to every event along it."
- [x] "MIT-licensed".
- [x] "end-to-end".
- [x] Typos: I read the whole paper again. Besides making the spelling of "labelled" consistent, I fixed a comma splice in State of the field and a subject-verb disagreement, and reworded a few sentences where the shortening had muddled the sense.

While checking the rendered PDF I found two reference entries that printed badly, and fixed those too. The Aiken 2020 entry set both `publisher` and `journal` to Substack, so it came out as "In Substack. Substack."; the duplicate field is gone. The Savitzky entry had a stray period inside the author field, which printed as "Savitzky, Abraham., & Golay"; it now abbreviates correctly.

These are changes to the text only, with nothing touched in the software, so v1.1.1 and the archive at 10.5281/zenodo.21906346 should still stand and no re-archive is needed. Tell me if you would rather have a fresh tag anyway.
